# apps/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@boxlite.com
    name: BoxLite Platforms Inc.
    url: https://www.boxlite.io
  description: BoxLite AI platform API Docs
  title: BoxLite
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /organizations/invitations:
    get:
      operationId: listOrganizationInvitationsForAuthenticatedUser
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/OrganizationInvitation"
                type: array
          description: List of organization invitations
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List organization invitations for authenticated user
      tags:
      - organizations
  /organizations/invitations/count:
    get:
      operationId: getOrganizationInvitationsCountForAuthenticatedUser
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: number
          description: Count of organization invitations
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get count of organization invitations for authenticated user
      tags:
      - organizations
  /organizations/invitations/{invitationId}/accept:
    post:
      operationId: acceptOrganizationInvitation
      parameters:
      - description: Invitation ID
        explode: false
        in: path
        name: invitationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationInvitation"
          description: Organization invitation accepted successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Accept organization invitation
      tags:
      - organizations
  /organizations/invitations/{invitationId}/decline:
    post:
      operationId: declineOrganizationInvitation
      parameters:
      - description: Invitation ID
        explode: false
        in: path
        name: invitationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Organization invitation declined successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Decline organization invitation
      tags:
      - organizations
  /organizations:
    get:
      operationId: listOrganizations
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Organization"
                type: array
          description: List of organizations
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List organizations
      tags:
      - organizations
    post:
      operationId: createOrganization
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganization"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
          description: Organization created successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Create organization
      tags:
      - organizations
  /organizations/{organizationId}/name:
    patch:
      operationId: updateOrganizationName
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationName"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
          description: Organization name updated successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update organization name
      tags:
      - organizations
  /organizations/{organizationId}/default-region:
    patch:
      operationId: setOrganizationDefaultRegion
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationDefaultRegion"
        required: true
      responses:
        "204":
          description: Default region set successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Set default region for organization
      tags:
      - organizations
  /organizations/{organizationId}:
    delete:
      operationId: deleteOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Organization deleted successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Delete organization
      tags:
      - organizations
    get:
      operationId: getOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
          description: Organization details
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get organization by ID
      tags:
      - organizations
  /organizations/{organizationId}/leave:
    post:
      operationId: leaveOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Organization left successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Leave organization
      tags:
      - organizations
  /organizations/{organizationId}/suspend:
    post:
      operationId: suspendOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationSuspension"
        required: false
      responses:
        "204":
          description: Organization suspended successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Suspend organization
      tags:
      - organizations
  /organizations/{organizationId}/unsuspend:
    post:
      operationId: unsuspendOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Organization unsuspended successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Unsuspend organization
      tags:
      - organizations
  /organizations/by-box-id/{boxId}:
    get:
      operationId: getOrganizationByBoxId
      parameters:
      - description: Box ID
        explode: false
        in: path
        name: boxId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
          description: Organization
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get organization by box ID
      tags:
      - organizations
  /organizations/otel-config/by-box-auth-token/{authToken}:
    get:
      operationId: getOrganizationOtelConfigByBoxAuthToken
      parameters:
      - description: Box Auth Token
        explode: false
        in: path
        name: authToken
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtelConfig"
          description: OTEL Config
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get organization OTEL config by box auth token
      tags:
      - organizations
  /organizations/{organizationId}/box-default-limited-network-egress:
    post:
      operationId: updateBoxDefaultLimitedNetworkEgress
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationBoxDefaultLimitedNetworkEgress"
        required: true
      responses:
        "204":
          description: Box default limited network egress updated successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update box default limited network egress
      tags:
      - organizations
  /organizations/{organizationId}/experimental-config:
    put:
      operationId: updateExperimentalConfig
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              example:
                otel:
                  endpoint: http://otel-collector:4317
                  headers:
                    api-key: XXX
              type: object
        description: Experimental configuration as a JSON object. Set to null to clear
          the configuration.
        required: false
      responses:
        "200":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update experimental configuration
      tags:
      - organizations
...
  /organizations/{organizationId}/invitations:
    get:
      operationId: listOrganizationInvitations
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/OrganizationInvitation"
                type: array
          description: List of pending organization invitations
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List pending organization invitations
      tags:
      - organizations
    post:
      operationId: createOrganizationInvitation
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationInvitation"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationInvitation"
          description: Organization invitation created successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Create organization invitation
      tags:
      - organizations
...
components:
  schemas:
...
    OrganizationInvitation:
      example:
        id: id
        email: email
        invitedBy: invitedBy
        organizationId: organizationId
        organizationName: organizationName
        expiresAt: 2000-01-23T04:56:07.000+00:00
        status: pending
        role: owner
        assignedRoles:
        - id: id
          name: name
          description: description
          permissions:
          - write:registries
          - write:registries
          isGlobal: true
          createdAt: 2000-01-23T04:56:07.000+00:00
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - id: id
          name: name
          description: description
          permissions:
          - write:registries
          - write:registries
          isGlobal: true
          createdAt: 2000-01-23T04:56:07.000+00:00
          updatedAt: 2000-01-23T04:56:07.000+00:00
        createdAt: 2000-01-23T04:56:07.000+00:00
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Invitation ID
          type: string
        email:
          description: Email address of the invitee
          type: string
        invitedBy:
          description: Email address of the inviter
          type: string
        organizationId:
          description: Organization ID
          type: string
        organizationName:
          description: Organization name
          type: string
        expiresAt:
          description: Expiration date of the invitation
          format: date-time
          type: string
        status:
          description: Invitation status
          enum:
          - pending
          - accepted
          - declined
          - cancelled
          type: string
        role:
          description: Member role
          enum:
          - owner
          - member
          type: string
        assignedRoles:
          description: Assigned roles
          items:
            $ref: "#/components/schemas/OrganizationRole"
          type: array
        createdAt:
          description: Creation timestamp
          format: date-time
          type: string
        updatedAt:
          description: Last update timestamp
          format: date-time
          type: string
      required:
      - assignedRoles
      - createdAt
      - email
      - expiresAt
      - id
      - invitedBy
      - organizationId
      - organizationName
      - role
      - status
      - updatedAt
      type: object
...
    UpdateOrganizationName:
      example:
        name: Default Organization
      properties:
        name:
          description: The public name of the organization
          example: Default Organization
          type: string
      required:
      - name
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect